tetragon: Move generic_output to generic functions
Moving generic_output to generic functions and change it to use
process_call_heap in generic functions directly, so we don't need
to pass it as argument.

Passing map pointer through argument might end up storing and loading
it to/from heap which makes it scalar value and breaks verifier.
