+++ bpf/process/bpf_generic_tracepoint.c
@@ -266,7 +266,7 @@ generic_tracepoint_actions(void *ctx)
 __attribute__((section("tracepoint"), used)) int
 generic_tracepoint_output(void *ctx)
 {
-	return generic_output(ctx, (struct bpf_map_def *)&process_call_heap, MSG_OP_GENERIC_TRACEPOINT);
+	return generic_output(ctx, MSG_OP_GENERIC_TRACEPOINT);
 }
 
 char _license[] __attribute__((section("license"), used)) = "Dual BSD/GPL";
